+++ EatIT.Core/Entities/Rating.cs
@@ -13,7 +13,7 @@ namespace EatIT.Core.Entities
         public int Id { get; set; }
         public int UserId { get; set; }
         public int RestaurantId { get; set; }
-        public int Star { get; set; }
+        public double Star { get; set; }
         public string? Comment { get; set; }
         public DateTime CreateAt { get; set; }
